requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
websockets>=13,<16
google-genai>=1.0
python-dotenv>=1.0
aic-sdk>=0.7
numpy>=1.26
pipecat-ai[websocket,gradium]>=1.0,<2.0
elevenlabs>=2.44,<3.0
reportlab>=4.2,<5.0
